Add Flyway DB migration plugin and configuration
The commit introduces the addition of the Flyway DB migration plugin in the build.gradle.kts file for version control of the database. It also includes a DataSource data class for setting up the connection to the database using environment variables or project properties. The Flyway configuration makes use of this data class properties.

Alongside this, a new gradle.properties file is also added, which contains the default properties required for the gradle build script. These values could either be overwritten by setting environment variables or directly in the script.

This change facilitates easier and better-managed database migrations across different environments of the application.